Moving import and making sure docstring pytest works
