+++ src/scores/plotdata/__init__.py
@@ -4,7 +4,16 @@ Import the functions from the implementations into the public API
 
 from scores.plotdata.murphy_impl import murphy_score, murphy_thetas
 from scores.plotdata.qq_impl import qq
+from scores.plotdata.rev_impl import relative_economic_value, relative_economic_value_from_rates
 from scores.plotdata.roc_impl import roc
 from scores.probability.rank_hist_impl import rank_histogram
 
-__all__ = ["murphy_score", "murphy_thetas", "qq", "roc", "rank_histogram"]
+__all__ = [
+    "murphy_score",
+    "murphy_thetas",
+    "qq",
+    "relative_economic_value",
+    "relative_economic_value_from_rates",
+    "roc",
+    "rank_histogram",
+]
