+++ tests/plotdata/test_rev.py
@@ -1,5 +1,5 @@
 """
-Contains unit tests for scores.probability.rev_impl
+Contains unit tests for scores.plotdata.rev_impl
 """
 
 import re
@@ -9,18 +9,18 @@ import numpy as np
 import pytest
 import xarray as xr
 
-from scores.probability import (
-    relative_economic_value,
-    relative_economic_value_from_rates,
-)
-from scores.probability.rev_impl import (
+from scores.plotdata.rev_impl import (
     _calculate_rev_core,
     _create_output_dataset,
     calculate_climatology,
     check_monotonic_array,
 )
+from scores.probability import (
+    relative_economic_value,
+    relative_economic_value_from_rates,
+)
 from scores.utils import ERROR_INVALID_WEIGHTS
-from tests.probability import rev_test_data as rtd
+from tests.plotdata import rev_test_data as rtd
 
 
 @pytest.fixture(name="make_contingency_data")
